feat: move copperAuth to own file
